requirements.txt
numpy==1.14.1
#falcon==1.2.0
#inflect==0.3.1
#audioread==2.1.5
librosa==0.5.1
matplotlib==2.0.2
#scipy==1.0.0
tqdm==4.11.2
#Unidecode==0.4.20
#pyaudio==0.2.11
#sounddevice==0.3.10
tensorflow-gpu==1.12.0
tensorboardX
Pillow
ipykernel
ipywidgets
mir_eval
